feat: add drawing queries(#109)
